svn merge -r882818:890596 http://svn.apache.org/repos/asf/hadoop/pig/trunk
git-svn-id: https://svn.apache.org/repos/asf/hadoop/pig/branches/load-store-redesign@891828 13f79535-47bb-0310-9956-ffa450edef68

+++ src/org/apache/pig/backend/hadoop/executionengine/mapReduceLayer/partitioners/SkewedPartitioner.java
@@ -53,10 +53,10 @@ public class SkewedPartitioner extends Partitioner<PigNullableWritable, Writable
     @Override
     public int getPartition(PigNullableWritable wrappedKey, Writable value,
             int numPartitions) {
-        // for streaming tables, return the partition index blindly
-        if (wrappedKey instanceof NullablePartitionWritable && ((int)((NullablePartitionWritable)wrappedKey).getPartition()) != -1) {
-            return (int) ((NullablePartitionWritable)wrappedKey).getPartition();
-        }
+		// for streaming tables, return the partition index blindly
+		if (wrappedKey instanceof NullablePartitionWritable && (((NullablePartitionWritable)wrappedKey).getPartition()) != -1) {
+			return ((NullablePartitionWritable)wrappedKey).getPartition();
+		}
 
         // for partition table, compute the index based on the sampler output
         Pair <Integer, Integer> indexes;
